requirements.txt
python-calendarific
openai
azure-identity
requests
beautifulsoup4
termcolor